+++ api/pkg/fjarm/helloworld/helloworldv1/BUILD.bazel
@@ -1,0 +1,18 @@
+load("@rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "helloworldv1",
+    srcs = ["helloworld.go"],
+    importpath = "github.com/fjarm/fjarm/api/pkg/fjarm/helloworld/helloworldv1",
+    visibility = ["//visibility:public"],
+)
+
+go_test(
+    name = "helloworldv1_test",
+    srcs = ["helloworld_test.go"],
+    embed = [":helloworldv1"],
+    deps = [
+        "@build_buf_gen_go_fjarm_fjarm_protocolbuffers_go//fjarm/helloworld/v1:helloworld",
+        "@com_github_bufbuild_protovalidate_go//:protovalidate-go",
+    ],
+)
