--- api/pkg/fjarm/helloworld/v1/BUILD.bazel
@@ -1,19 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library", "go_test")
-
-go_test(
-    name = "helloworld_test",
-    srcs = ["helloworld_test.go"],
-    embed = [":helloworld"],
-    importpath = "github.com/fjarm/fjarm/api/pkg/fjarm/helloworld/v1",
-    deps = [
-        "@build_buf_gen_go_fjarm_fjarm_protocolbuffers_go//fjarm/helloworld/v1:helloworld",
-        "@com_github_bufbuild_protovalidate_go//:protovalidate-go",
-    ],
-)
-
-go_library(
-    name = "helloworld",
-    srcs = ["helloworld.go"],
-    importpath = "github.com/fjarm/fjarm/api/pkg/fjarm/helloworld/v1",
-    visibility = ["//visibility:public"],
-)
