--- api/pkg/fjarm/helloworld/v1/helloworld.go
@@ -1,1 +1,0 @@
-package v1
